Add symfony/console command to test the idea
This change introduce a symfony/console component which makes
work with console interface easier. To show how it works,
I implement an `analyse` command which use the current code.
If we decided to move in this direction, we should reconsider our
approach to handling parameter and commands as well as isolating
the core code.